requirements.txt
# ==============================================================================
# DEPENDENCIAS DEL PROYECTO AJEDREZ v2.0
# ==============================================================================
# Versiones probadas con Python 3.14.2 en Windows 10/11
# Instalación: pip install -r requirements.txt
# ==============================================================================

# MOTOR GRÁFICO Y MULTIMEDIA
# pygame-ce (Community Edition) - Último soporte para Python 3.14+
# Incluye: Renderización 2D, manejo de eventos, sprites, fuentes, sonido
pygame-ce==2.5.6

# MOTORES Y LIBRERÍAS DE AJEDREZ
# python-chess - Validación de movimientos, generación de FEN, integración UCI
python-chess==1.999

# ==============================================================================
# LIBRERÍAS ESTÁNDAR (No requieren instalación vía pip)
# ==============================================================================
# socket       - Comunicación TCP/UDP para modo LAN (servidores y clientes)
# json         - Protocolo de serialización para mensajes entre jugadores
# threading    - Hilos asincronos para escucha de movimientos en red
# typing       - Anotaciones de tipos (Python 3.5+)
# enum         - Enumeraciones: Color, TipoPieza, EstadoJuego
# os           - Rutas relativas, gestión de directorios
# subprocess   - Ejecución de motores UCI externos (Stockfish, LCZero)
# pickle       - Serialización de objetos (historial de partidas)

# ==============================================================================
# LIBRERÍAS OPCIONALES (Comentadas por defecto)
# ==============================================================================
# Descomenta si quieres agregar funcionalidades futuras:

# requests>=2.31.0     - API Chess.com, online ratings, bases de datos
# pillow>=10.0.0       - Procesamiento de imágenes para tablero personalizado
# numpy>=1.24.0        - Cálculos numéricos avanzados para IA
# pyaudio>=0.2.13      - Captura de audio (futuro: comandos de voz)
# websockets>=12.0     - Protocolo WebSocket para servidores online

